requirements.txt
streamlit==1.31.0
opencv-python==4.9.0.80
numpy==1.26.4
pillow==10.2.0
insightface==0.7.3
onnxruntime==1.17.1
scikit-learn==1.4.0
pandas==2.2.0
pyyaml==6.0.1
sqlalchemy==2.0.25
loguru==0.7.2
